src/face_res/requirements.txt
matplotlib==3.3.4
numpy==1.19.5
opencv_python==4.5.4.60
Pillow==8.4.0
scikit_image==0.17.2
torch==1.10.0
torchvision==0.11.1
flask==2.0.2
requests==2.26.0
basicsr>=1.3.3.11
moviepy==1.0.3